requirements.txt
# Flask Web Framework
flask
flask-cors

# Geospatial Data Access
pystac-client
odc-stac
planetary-computer

# Data Processing
xarray
rioxarray
numpy

# Geospatial Libraries
geopandas
geopy
rasterio

# Visualization & Image Generation
matplotlib
pillow
